requirements.txt
Django==2.1.2
djangorestframework==3.9.0
py==1.7.0
pytest==3.9.3
psycopg2==2.7.5
pytest-django==3.4.3
